+++ cmd/github.go
@@ -2,7 +2,7 @@
 // Created by make-go-github-file.yaml on Sun Jun 18 19:44:32 UTC 2023
 package cmd
 
-var githubDate string = "2023-06-18"
+var githubDate string = "2023-07-21"
 var githubIteration string = "0"
 var githubRef string = "refs/tags/0.2.7"
 var githubRefName string = "0.2.7"